+++ lib/ui/page/profile/follow/followerListPage.dart
@@ -1,6 +1,10 @@
+import 'package:flutter_twitter_clone/ui/theme/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_twitter_clone/model/user.dart';
 import 'package:flutter_twitter_clone/ui/page/common/usersListPage.dart';
+import 'package:flutter_twitter_clone/ui/page/profile/follow/followListState.dart';
+import 'package:flutter_twitter_clone/widgets/newWidget/customLoader.dart';
+import 'package:provider/provider.dart';
 
 class FollowerListPage extends StatelessWidget {
   const FollowerListPage({Key? key, this.userList, this.profile})
@@ -12,9 +16,9 @@ class FollowerListPage extends StatelessWidget {
       {required List<String> userList, required UserModel profile}) {
     return MaterialPageRoute(
       builder: (BuildContext context) {
-        return FollowerListPage(
-          profile: profile,
-          userList: userList,
+        return ChangeNotifierProvider(
+          create: (_) => FollowListState(StateType.follower),
+          child: FollowerListPage(userList: userList, profile: profile),
         );
       },
     );
@@ -22,6 +26,16 @@ class FollowerListPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    if (context.watch<FollowListState>().isbusy) {
+      return SizedBox(
+        height: context.height,
+        child: CustomScreenLoader(
+          height: double.infinity,
+          width: context.width,
+          backgroundColor: Colors.white,
+        ),
+      );
+    }
     return UsersListPage(
       pageTitle: 'Followers',
       userIdsList: userList,
@@ -29,6 +43,12 @@ class FollowerListPage extends StatelessWidget {
       emptyScreenText: '${profile?.userName} doesn\'t have any followers',
       emptyScreenSubTileText:
           'When someone follow them, they\'ll be listed here.',
+      isFollowing: (user) {
+        return context.watch<FollowListState>().isFollowing(user);
+      },
+      onFollowPressed: (user) {
+        context.read<FollowListState>().followUser(user);
+      },
     );
   }
 }