+++ lib/widgets/url_text/link_preview.dart
@@ -1,13 +1,12 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:link_preview_generator/link_preview_generator.dart';
-
 import 'package:flutter_twitter_clone/helper/utility.dart';
 import 'package:flutter_twitter_clone/state/feedState.dart';
 import 'package:flutter_twitter_clone/ui/theme/theme.dart';
+import 'package:link_preview_generator/link_preview_generator.dart';
+// ignore: implementation_imports
 import 'package:link_preview_generator/src/utils/analyzer.dart'
-    show
-        LinkPreviewAnalyzer; //FIXME Don't import implementation files from another package.
+    show LinkPreviewAnalyzer;
 import 'package:provider/provider.dart';
 
 class LinkPreviewer extends StatelessWidget {